requirements.txt
Django
dj-rest-auth
django-allauth
django-cors-headers
djangorestframework
drf-spectacular
gunicorn
whitenoise
